Decouple table from settings view component to reuse in another component
